+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/actions/FlexibleSearchToggleParametersEditorAction.kt
@@ -34,14 +34,10 @@ class FlexibleSearchToggleParametersEditorAction : ToggleAction(
 
     override fun getActionUpdateThread() = ActionUpdateThread.BGT
 
-    override fun isSelected(e: AnActionEvent): Boolean = e.flexibleSearchSplitEditor()?.isParametersPanelVisible()
+    override fun isSelected(e: AnActionEvent): Boolean = e.flexibleSearchSplitEditor()?.inEditorParameters
         ?: false
 
     override fun setSelected(e: AnActionEvent, state: Boolean) {
-        e.flexibleSearchSplitEditor()
-            ?.apply {
-                if (state) showParametersPanel()
-                else hideParametersPanel()
-            }
+        e.flexibleSearchSplitEditor()?.inEditorParameters = state
     }
 }