+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/editor/FlexibleSearchSplitEditor.kt
@@ -18,56 +18,33 @@
 
 package com.intellij.idea.plugin.hybris.flexibleSearch.editor
 
-import com.intellij.database.editor.CsvTableFileEditor
-import com.intellij.idea.plugin.hybris.flexibleSearch.FlexibleSearchLanguage
-import com.intellij.idea.plugin.hybris.flexibleSearch.psi.FlexibleSearchBindParameter
-import com.intellij.idea.plugin.hybris.grid.GridXSVFormatService
+import com.intellij.idea.plugin.hybris.flexibleSearch.editor.FlexibleSearchInEditorParametersView.buildParametersPanel
 import com.intellij.idea.plugin.hybris.system.meta.MetaModelChangeListener
 import com.intellij.idea.plugin.hybris.system.meta.MetaModelStateService
 import com.intellij.idea.plugin.hybris.system.type.meta.TSGlobalMetaModel
-import com.intellij.idea.plugin.hybris.system.type.meta.TSMetaModelStateService
 import com.intellij.idea.plugin.hybris.tools.remote.http.impex.HybrisHttpResult
-import com.intellij.idea.plugin.hybris.ui.Dsl
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.PlatformDataKeys
 import com.intellij.openapi.application.edtWriteAction
-import com.intellij.openapi.components.service
 import com.intellij.openapi.fileEditor.FileEditor
 import com.intellij.openapi.fileEditor.FileEditorState
 import com.intellij.openapi.fileEditor.TextEditor
-import com.intellij.openapi.fileTypes.PlainTextFileType
-import com.intellij.openapi.project.DumbService
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.getPreferredFocusedComponent
 import com.intellij.openapi.util.Disposer
 import com.intellij.openapi.util.Key
 import com.intellij.openapi.util.UserDataHolderBase
 import com.intellij.openapi.util.getOrCreateUserData
-import com.intellij.openapi.util.text.StringUtil
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.pom.Navigatable
 import com.intellij.psi.PsiDocumentManager
-import com.intellij.psi.tree.IElementType
-import com.intellij.psi.util.PsiTreeUtil
-import com.intellij.testFramework.LightVirtualFile
-import com.intellij.ui.AnimatedIcon
-import com.intellij.ui.EditorNotificationPanel
-import com.intellij.ui.InlineBanner
 import com.intellij.ui.OnePixelSplitter
-import com.intellij.ui.dsl.builder.*
-import com.intellij.ui.dsl.gridLayout.UnscaledGaps
-import com.intellij.util.application
 import com.intellij.util.asSafely
-import com.intellij.util.ui.JBUI
-import com.michaelbaranov.microba.calendar.DatePicker
 import kotlinx.coroutines.*
 import java.awt.BorderLayout
-import java.awt.Dimension
 import java.beans.PropertyChangeListener
 import java.io.Serial
-import java.text.SimpleDateFormat
-import java.util.*
 import javax.swing.JComponent
 import javax.swing.JPanel
 import kotlin.time.Duration
@@ -76,11 +53,73 @@ import kotlin.time.Duration.Companion.milliseconds
 fun AnActionEvent.flexibleSearchSplitEditor() = this.getData(PlatformDataKeys.FILE_EDITOR)
     ?.asSafely<FlexibleSearchSplitEditor>()
 
-class FlexibleSearchSplitEditor(private val textEditor: TextEditor, private val project: Project) : UserDataHolderBase(), FileEditor, TextEditor {
+class FlexibleSearchSplitEditor(internal val textEditor: TextEditor, private val project: Project) : UserDataHolderBase(), FileEditor, TextEditor {
+
+    companion object {
+        @Serial
+        private const val serialVersionUID: Long = -3770395176190649196L
+        internal val KEY_FLEXIBLE_SEARCH_PARAMETERS: Key<Collection<FlexibleSearchQueryParameter>> = Key.create("flexibleSearch.parameters.key")
+        private val KEY_IN_EDITOR_RESULTS: Key<Boolean> = Key.create("flexibleSearch.in_editor_results.key")
+    }
+
+    var inEditorParameters: Boolean
+        get() = inEditorParametersView != null
+        set(state) {
+            if (state) {
+                with(buildParametersPanel(project, this)) {
+                    inEditorParametersView = this
+                }
+            } else {
+                queryParametersDisposable?.apply { Disposer.dispose(this) }
+                queryParametersDisposable = null
+                inEditorParametersView = null
+            }
+
+            component.requestFocus()
+            horizontalSplitter.firstComponent.requestFocus()
+
+            reparseTextEditor()
+        }
+
+    val queryParameters: Collection<FlexibleSearchQueryParameter>?
+        get() = if (inEditorParameters) getUserData(KEY_FLEXIBLE_SEARCH_PARAMETERS)
+        else null
+
+    val query: String
+        get() = queryParameters
+            ?.sortedByDescending { it.name.length }
+            ?.let { properties ->
+                var updatedContent = getText()
+                properties.forEach {
+                    updatedContent = updatedContent.replace("?${it.name}", it.value)
+                }
+                return@let updatedContent
+            }
+            ?: getText()
+
+    var inEditorResults: Boolean
+        get() = getOrCreateUserData(KEY_IN_EDITOR_RESULTS) { false }
+        set(state) {
+            putUserData(KEY_IN_EDITOR_RESULTS, state)
+            verticalSplitter.secondComponent?.isVisible = state
+        }
+
+    internal var inEditorResultsView: JComponent?
+        get() = verticalSplitter.secondComponent
+        set(view) {
+            verticalSplitter.secondComponent = view
+        }
+
+    internal var inEditorParametersView: JComponent?
+        get() = horizontalSplitter.secondComponent
+        set(view) {
+            horizontalSplitter.secondComponent = view
+        }
+
+    internal var queryParametersDisposable: Disposable? = null
 
     private var renderParametersJob: Job? = null
-    private var refreshTextEditorJob: Job? = null
-    private var parametersPanelDisposable: Disposable? = null
+    private var reparseTextEditorJob: Job? = null
 
     private val horizontalSplitter = OnePixelSplitter(false).apply {
         isShowDividerControls = true
@@ -98,86 +137,44 @@ class FlexibleSearchSplitEditor(private val textEditor: TextEditor, private val 
         firstComponent = horizontalSplitter
     }
 
-    private val splitPanel = JPanel(BorderLayout()).apply {
+    private val rootPanel = JPanel(BorderLayout()).apply {
         add(verticalSplitter, BorderLayout.CENTER)
     }
-
-    var inEditorResults: Boolean
-        set(state) {
-            putUserData(KEY_IN_EDITOR_RESULTS, state)
-            verticalSplitter.secondComponent?.isVisible = state
-        }
-        get() = getOrCreateUserData(KEY_IN_EDITOR_RESULTS) { false }
 
     init {
         with(project.messageBus.connect(this)) {
             subscribe(MetaModelStateService.TOPIC, object : MetaModelChangeListener {
                 override fun typeSystemChanged(globalMetaModel: TSGlobalMetaModel) {
                     refreshParameters()
-                    refreshTextEditor()
+                    reparseTextEditor()
                 }
             })
         }
     }
 
-    fun showExecutionResult(result: HybrisHttpResult) {
-        if (result.errorMessage.isNotBlank()) {
-            verticalSplitter.secondComponent = executionResultsErrorPane(result)
-        } else {
-            executionResultsPane(result)
-        }
-    }
+    fun renderExecutionResult(result: HybrisHttpResult) = FlexibleSearchInEditorResultsView
+        .renderExecutionResult(project, this, result)
 
-    fun pendingExecutionResult() {
-        if (verticalSplitter.secondComponent == null) return
+    fun beforeExecution() = FlexibleSearchInEditorResultsView
+        .renderRunningExecution(this)
 
-        verticalSplitter.secondComponent = panel {
-            panel {
-                row {
-                    cell(
-                        InlineBanner(
-                            "Executing HTTP Call to SAP Commerce...",
-                            EditorNotificationPanel.Status.Info
-                        )
-                            .showCloseButton(false)
-                            .setIcon(AnimatedIcon.Default.INSTANCE)
-                    )
-                        .align(Align.FILL)
-                        .resizableColumn()
-                }.topGap(TopGap.SMALL)
-            }
-                .customize(UnscaledGaps(16, 16, 16, 16))
-        }
-    }
-
-    fun getParameters() = if (isParametersPanelVisible()) getUserData(KEY_FLEXIBLE_SEARCH_PARAMETERS)
-    else null
-
-    fun getQuery(): String = getParameters()
-        ?.sortedByDescending { it.name.length }
-        ?.let { properties ->
-            var updatedContent = getText()
-            properties.forEach {
-                updatedContent = updatedContent.replace("?${it.name}", it.value)
-            }
-            return@let updatedContent
-        }
-        ?: getText()
-
-    fun refreshParameters(delayMs: Duration = 250.milliseconds) {
+    fun refreshParameters(delayMs: Duration = 500.milliseconds) {
         renderParametersJob?.cancel()
         renderParametersJob = CoroutineScope(Dispatchers.Default).launch {
             delay(delayMs)
 
-            if (project.isDisposed || !isParametersPanelVisible()) return@launch
+            if (project.isDisposed || !inEditorParameters) return@launch
 
-            horizontalSplitter.secondComponent = buildParametersPanel()
+            inEditorParametersView = buildParametersPanel(project, this@FlexibleSearchSplitEditor)
         }
     }
 
-    fun refreshTextEditor(delayMs: Duration = 1000.milliseconds) {
-        refreshTextEditorJob?.cancel()
-        refreshTextEditorJob = CoroutineScope(Dispatchers.Default).launch {
+    /**
+     * Reparse PsiFile in the related TextEditor to retrigger inline hints computation
+     */
+    internal fun reparseTextEditor(delayMs: Duration = 1000.milliseconds) {
+        reparseTextEditorJob?.cancel()
+        reparseTextEditorJob = CoroutineScope(Dispatchers.Default).launch {
             delay(delayMs)
 
             if (project.isDisposed) return@launch
@@ -185,168 +182,6 @@ class FlexibleSearchSplitEditor(private val textEditor: TextEditor, private val 
             edtWriteAction {
                 PsiDocumentManager.getInstance(project).reparseFiles(listOf(file), false)
             }
-        }
-    }
-
-    fun hideParametersPanel() {
-        parametersPanelDisposable?.apply { Disposer.dispose(this) }
-        parametersPanelDisposable = null
-        horizontalSplitter.secondComponent = null
-
-        component.requestFocus()
-        horizontalSplitter.firstComponent.requestFocus()
-
-        refreshTextEditor()
-    }
-
-    fun showParametersPanel() {
-        with(buildParametersPanel()) {
-            horizontalSplitter.secondComponent = this
-        }
-
-        component.requestFocus()
-        horizontalSplitter.firstComponent.requestFocus()
-
-        refreshTextEditor()
-    }
-
-    fun isParametersPanelVisible(): Boolean = horizontalSplitter.secondComponent != null
-
-    fun buildParametersPanel(): JComponent? {
-        if (project.isDisposed) return null
-
-        parametersPanelDisposable?.let { Disposer.dispose(it) }
-
-        val parentDisposable = Disposer.newDisposable().apply {
-            parametersPanelDisposable = this
-            Disposer.register(textEditor, this)
-        }
-
-        if (!isTypeSystemInitialized()) return panel {
-            row {
-                label("Initializing Type System, please wait...")
-                    .align(Align.CENTER)
-                    .resizableColumn()
-            }.resizableRow()
-        }
-
-        val parameters = collectParameters()
-
-        //extract to small methods: render headers, render no data panel, render data panel
-        return panel {
-            panel {
-                row {
-                    val infoBanner = InlineBanner(
-                        """
-                        <html><body style='width: 100%'>
-                        <p>This feature may be unstable. Use with caution.</p>
-                        <p>Submit issues or suggestions to project's GitHub repository.</p>
-                        </body></html>
-                    """.trimIndent(),
-                        EditorNotificationPanel.Status.Warning
-                    )
-
-                    cell(infoBanner)
-                        .align(Align.FILL)
-                        .resizableColumn()
-                }.topGap(TopGap.SMALL)
-            }
-                .customize(UnscaledGaps(16, 16, 16, 16))
-
-            //todo extract from panel to show message vertical center aligned
-            panel {
-                if (parameters.isEmpty()) {
-                    row {
-                        label("FlexibleSearch query doesn't have parameters")
-                            .align(Align.CENTER)
-                    }
-                } else {
-                    group("Parameters") {
-                        parameters.forEach { parameter ->
-                            row {
-                                //todo limit the long name depends on width of the panel
-                                // TODO: migrate to proper property binding
-                                when (parameter.type) {
-                                    "java.lang.Float", "java.lang.Double", "java.lang.Byte", "java.lang.Short", "java.lang.Long", "java.lang.Integer",
-                                    "float", "double", "byte", "short", "long", "int" -> intTextField()
-                                        .label("${parameter.name}:")
-                                        .align(AlignX.FILL)
-                                        .text(parameter.value)
-                                        .onChanged { applyValue(parameter, it.text) { it.text } }
-
-                                    "boolean",
-                                    "java.lang.Boolean" -> checkBox(parameter.name)
-                                        .align(AlignX.FILL)
-                                        .selected(parameter.value == "1")
-                                        .onChanged {
-                                            val presentationValue = if (it.isSelected) "true" else "false"
-                                            applyValue(parameter, presentationValue) { if (it.isSelected) "1" else "0" }
-                                        }
-                                        .also {
-                                            parameter.value = (if (parameter.value == "1") "1" else "0")
-                                        }
-
-                                    "java.util.Date" -> cell(
-                                        DatePicker(
-                                            parameter.value.toLongOrNull()?.let { Date(it) },
-                                            SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS")
-                                        )
-                                    )
-                                        .label("${parameter.name}:")
-                                        .align(Align.FILL).apply {
-                                            component.also { datePicker ->
-                                                val listener = PropertyChangeListener { event ->
-                                                    if (event.propertyName == "date") {
-                                                        val newValue = event.newValue?.asSafely<Date>()
-                                                        val presentationValue = newValue?.let { date -> SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS").format(date) } ?: ""
-
-                                                        applyValue(parameter, presentationValue) {
-                                                            newValue?.time?.toString() ?: ""
-                                                        }
-                                                    }
-                                                }
-                                                datePicker.addPropertyChangeListener(listener)
-                                                Disposer.register(parentDisposable) {
-                                                    datePicker.removePropertyChangeListener(listener)
-                                                }
-                                            }
-                                        }
-
-                                    "String",
-                                    "java.lang.String",
-                                    "localized:java.lang.String" -> textField()
-                                        .label("${parameter.name}:")
-                                        .align(AlignX.FILL)
-                                        .text(StringUtil.unquoteString(parameter.value, '\''))
-                                        .onChanged { applyValue(parameter, "'${it.text}'") { "'${it.text}'" } }
-
-                                    else -> textField()
-                                        .label("${parameter.name}:")
-                                        .align(AlignX.FILL)
-                                        .text(parameter.value)
-                                        .onChanged { applyValue(parameter, it.text) { "${it.text}" } }
-                                }
-
-                            }.layout(RowLayout.PARENT_GRID)
-                        }
-                    }
-                }
-            }
-        }
-            .apply { border = JBUI.Borders.empty(5, 16, 10, 16) }
-            .let { Dsl.scrollPanel(it) }
-            .apply {
-                minimumSize = Dimension(minimumSize.width, 165)
-            }
-    }
-
-    private fun applyValue(parameter: FlexibleSearchParameter, presentationValue: String, newValueProvider: () -> String) {
-        val originalValue = parameter.value
-        parameter.presentationValue = presentationValue
-        parameter.value = newValueProvider.invoke()
-
-        if (originalValue != parameter.value) {
-            refreshTextEditor()
         }
     }
 
@@ -363,7 +198,7 @@ class FlexibleSearchSplitEditor(private val textEditor: TextEditor, private val 
     override fun getPreferredFocusedComponent(): JComponent? = if (textEditor.component.isVisible) textEditor.preferredFocusedComponent
     else component.getPreferredFocusedComponent()
 
-    override fun getComponent() = splitPanel
+    override fun getComponent() = rootPanel
     override fun getName() = "FlexibleSearch Split Editor"
     override fun setState(state: FileEditorState) = textEditor.setState(state)
     override fun isModified() = textEditor.isModified
@@ -374,110 +209,7 @@ class FlexibleSearchSplitEditor(private val textEditor: TextEditor, private val 
     override fun navigateTo(navigatable: Navigatable) = textEditor.navigateTo(navigatable)
     override fun getFile(): VirtualFile? = editor.virtualFile
 
-    private fun isTypeSystemInitialized(): Boolean {
-        if (project.isDisposed) return false
-        if (DumbService.isDumb(project)) return false
-
-        try {
-            val metaModelStateService = project.service<TSMetaModelStateService>()
-            metaModelStateService.get()
-
-            return metaModelStateService.initialized()
-        } catch (_: Throwable) {
-            return false
-        }
-    }
-
-    private fun collectParameters(): Collection<FlexibleSearchParameter> {
-        val currentParameters = getUserData(KEY_FLEXIBLE_SEARCH_PARAMETERS) ?: emptySet()
-
-        val parameters = application.runReadAction<Collection<FlexibleSearchParameter>> {
-            PsiDocumentManager.getInstance(project).getPsiFile(editor.document)
-                ?.let { PsiTreeUtil.findChildrenOfType(it, FlexibleSearchBindParameter::class.java) }
-                ?.map { FlexibleSearchParameter.of(it, currentParameters) }
-                ?.distinctBy { it.name }
-                ?: emptySet()
-        }
-
-        putUserData(KEY_FLEXIBLE_SEARCH_PARAMETERS, parameters)
-        return parameters
-    }
-
     private fun getText(): String = editor.selectionModel.selectedText
         .takeIf { selectedText -> selectedText != null && selectedText.trim { it <= ' ' }.isNotEmpty() }
         ?: editor.document.text
-
-    private fun executionResultsPane(result: HybrisHttpResult) {
-        CoroutineScope(Dispatchers.Default).launch {
-            if (project.isDisposed) return@launch
-
-            val lvf = LightVirtualFile(
-                this@FlexibleSearchSplitEditor.file?.name + ".fxs.result.csv",
-                PlainTextFileType.INSTANCE,
-                result.output
-            )
-
-            val format = project.service<GridXSVFormatService>().getFormat(FlexibleSearchLanguage)
-
-            edtWriteAction {
-                val editor = CsvTableFileEditor(project, lvf, format);
-                this@FlexibleSearchSplitEditor.verticalSplitter.secondComponent = editor.component
-            }
-        }
-    }
-
-    private fun executionResultsErrorPane(result: HybrisHttpResult) = panel {
-        panel {
-            row {
-                cell(
-                    InlineBanner(
-                        result.errorMessage,
-                        EditorNotificationPanel.Status.Error
-                    ).showCloseButton(false)
-                )
-                    .align(Align.FILL)
-                    .resizableColumn()
-            }.topGap(TopGap.SMALL)
-        }
-            .customize(UnscaledGaps(16, 16, 16, 16))
-    }
-
-    companion object {
-        @Serial
-        private const val serialVersionUID: Long = -3770395176190649196L
-        private val KEY_FLEXIBLE_SEARCH_PARAMETERS: Key<Collection<FlexibleSearchParameter>> = Key.create("flexibleSearch.parameters.key")
-        private val KEY_IN_EDITOR_RESULTS: Key<Boolean> = Key.create("flexibleSearch.in_editor_results.key")
-    }
-}
-
-data class FlexibleSearchParameter(
-    val name: String,
-    var value: String = "",
-    var presentationValue: String = value,
-    val type: String? = null,
-    val operand: IElementType? = null
-) {
-    companion object {
-        fun of(bindParameter: FlexibleSearchBindParameter, currentParameters: Collection<FlexibleSearchParameter>): FlexibleSearchParameter {
-            val parameter = bindParameter.value
-            val currentParameter = currentParameters.find { it.name == parameter }
-            val itemType = bindParameter.itemType
-            val value = currentParameter?.value ?: resolveInitialValue(itemType)
-            val presentationValue = currentParameter?.presentationValue ?: resolveInitialPresentationValue(itemType)
-
-            return FlexibleSearchParameter(parameter, value, presentationValue, type = itemType)
-        }
-
-        private fun resolveInitialValue(itemType: String?): String = when (itemType) {
-            "boolean", "java.lang.Boolean" -> "0"
-            "String", "java.lang.String", "localized:java.lang.String" -> "''"
-            else -> ""
-        }
-
-        private fun resolveInitialPresentationValue(itemType: String?): String = when (itemType) {
-            "boolean", "java.lang.Boolean" -> "false"
-            "String", "java.lang.String", "localized:java.lang.String" -> "''"
-            else -> ""
-        }
-    }
 }